lambda/predict/requirements.txt
scikit-learn
pandas
mysql-connector-python
boto3
wandb
python-dotenv
